+++ variables.tf
@@ -45,12 +45,25 @@ variable "custom_origins" {
     https_port               = number
     custom_headers           = list(any)
   }))
+  default = []
+}
+variable "s3_origins" {
+  type = list(object({
+    domain_name            = string
+    origin_id              = string
+    origin_access_identity = string
+  }))
+  default = []
 }
 variable "geo_restriction" {
   type = object({
     type      = string
     locations = list(string)
   })
+  default = {
+    type      = "none"
+    locations = []
+  }
 }
 variable "price_class" {
   type    = string
@@ -70,6 +83,7 @@ variable "default_cache_behavior" {
     forward_headers                = list(string)
     forward_querystring            = bool
     forward_querystring_cache_keys = list(string)
+    lambda_function_association    = list(any)
   })
 }
 variable "cache_behaviors" {
@@ -87,5 +101,12 @@ variable "cache_behaviors" {
     forward_headers                = list(string)
     forward_querystring            = bool
     forward_querystring_cache_keys = list(string)
+    lambda_function_association    = list(any)
   }))
+  default = []
+}
+
+variable "default_root_object" {
+  type    = string
+  default = null
 }